# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Food Palace Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List current user's orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/confirm-payment:
    post:
      operationId: confirmPayment
      tags: [orders]
      summary: Customer confirms bank transfer made
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment confirmed by customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # FAVORITES
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: Admin list all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: paymentStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/orders/{id}/payment-status:
    patch:
      operationId: updatePaymentStatus
      tags: [admin]
      summary: Update payment status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # RESTAURANT SETTINGS
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [id, productName, quantity, unitPrice, subtotal]
      properties:
        id:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        variantName:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        addonNames:
          type: array
          items:
            type: string
        specialInstructions:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, userId, status, paymentMethod, paymentStatus, subtotal, deliveryFee, total, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerName:
          type: ["string", "null"]
        status:
          type: string
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        deliveryAddress:
          type: ["string", "null"]
        deliveryZoneId:
          type: ["integer", "null"]
        deliveryZoneName:
          type: ["string", "null"]
        deliveryNotes:
          type: ["string", "null"]
        estimatedDeliveryMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderInput:
      type: object
      required: [paymentMethod, deliveryAddress, deliveryZoneId]
      properties:
        paymentMethod:
          type: string
        deliveryAddress:
          type: string
        deliveryZoneId:
          type: integer
        deliveryNotes:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    PaymentStatusUpdate:
      type: object
      required: [paymentStatus]
      properties:
        paymentStatus:
          type: string

    # RESTAURANT SETTINGS
...