# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Palace Restaurant API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  customerName?: string | null;
  status: string;
  paymentMethod: string;
  paymentStatus: string;
  subtotal: number;
  deliveryFee: number;
  total: number;
  /** @nullable */
  deliveryAddress?: string | null;
  /** @nullable */
  deliveryZoneId?: number | null;
  /** @nullable */
  deliveryZoneName?: string | null;
  /** @nullable */
  deliveryNotes?: string | null;
  /** @nullable */
  estimatedDeliveryMinutes?: number | null;
  createdAt: string;
  items?: OrderItem[];
}