# lib/api-zod/src/generated/types/paymentStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food Palace Restaurant API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentStatusUpdate {
  paymentStatus: string;
}